@@ -3,36 +3,36 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "bootstrap": "~4.6.1",
-    "jquery": "~3.6.0",
-    "connected-react-router": "6.5.2",
-    "history": "4.10.1",
-    "merge": "2.1.1",
-    "popper.js": "^1.16.1",
-    "react": "~16.14.0",
-    "react-dom": "~16.14.0",
-    "react-redux": "~7.2.6",
-    "react-router": "~5.2.1",
-    "react-router-dom": "~5.3.0",
-    "reactstrap": "~8.10.1",
-    "redux": "4.1.2",
-    "redux-thunk": "2.4.1"
+    "@emotion/react": "^11.9.0",
+    "@emotion/styled": "^11.8.1",
+    "@mui/icons-material": "^5.6.2",
+    "@mui/material": "^5.6.4",
+    "@reduxjs/toolkit": "^1.8.1",
+    "@types/node": "^16.11.26",
+    "@types/react": "^17.0.44",
+    "@types/react-redux": "^7.1.23",
+    "@types/react-router-dom": "^5.3.3",
+    "bootstrap": "^5.1.3",
+    "jwt-decode": "^3.1.2",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
+    "react-redux": "^7.2.8",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "5.0.0",
+    "typescript": "^4.6.3",
+    "web-vitals": "^2.1.4"
   },
   "devDependencies": {
-    "@types/history": "4.7.10",
-    "@types/jest": "24.0.19",
-    "@types/node": "12.11.6",
-    "@types/react": "~16.9.14",
-    "@types/react-dom": "~16.9.14",
-    "@types/react-redux": "~7.1.22",
-    "@types/react-router": "~5.1.17",
-    "@types/react-router-dom": "~5.3.0",
-    "@types/reactstrap": "~8.7.2",
-    "cross-env": "6.0.3",
-    "eslint-plugin-jsx-a11y": "^6.4.1",
-    "nan": "^2.15.0",
-    "react-scripts": "^5.0.0",
-    "typescript": "3.6.4"
+    "@types/react-dom": "^18.0.3",
+    "@typescript-eslint/eslint-plugin": "^5.32.0",
+    "@typescript-eslint/parser": "^5.32.0",
+    "eslint": "^8.21.0",
+    "eslint-config-google": "^0.14.0",
+    "eslint-config-standard": "^17.0.0",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-n": "^15.2.4",
+    "eslint-plugin-promise": "^6.0.0",
+    "eslint-plugin-react": "^7.30.1"
   },
   "scripts": {
     "start": "react-scripts start",
