@@ -1,10 +1,22 @@
-import * as React from 'react';
-import * as ReactDOM from 'react-dom';
+import React from 'react';
+import {createRoot} from 'react-dom/client';
+import {Provider} from 'react-redux';
+import {BrowserRouter} from 'react-router-dom';
+import {store} from './store';
 import App from './App';
+import {firstResetAuthState} from './store/AuthReducer/AuthActionCreators';
 import registerServiceWorker from './registerServiceWorker';
 
-ReactDOM.render(
-    <App/>,
-    document.getElementById('root'));
+const container = document.getElementById('root');
+const root = createRoot(container!);
+
+store.dispatch(firstResetAuthState());
+root.render(
+    <Provider store={store}>
+      <BrowserRouter>
+        <App/>
+      </BrowserRouter>
+    </Provider>,
+);
 
 registerServiceWorker();
